tests/wpt/tests/css/css-conditional/js/CSS-supports-details-content-pseudo-parsing.html

<!doctype html>
<meta charset="utf-8">
<title>CSS.supports() ::details-content parsing</title>
<link rel="author" title="Luke Warlow" href="mailto:lwarlow@igalia.com">
<link rel="help" href="https://www.w3.org/TR/css-conditional-4/#the-css-namespace">
<link rel="help" href="https://drafts.csswg.org/css-pseudo-4/#element-like">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  test(function() {
    assert_equals(CSS.supports("selector(::details-content)"), true);
  }, "selector() function accepts ::details-content");

  test(function() {
    assert_equals(CSS.supports("selector(::details-content::before)"), true);
  }, "selector() function accepts ::details-content followed by ::before");

  test(function() {
    assert_equals(CSS.supports("selector(::details-content::first-line)"), true);
  }, "selector() function accepts ::details-content followed by ::first-line");

  test(function() {
    assert_equals(CSS.supports("selector(::details-content:hover"), true);
  }, "selector() function accepts ::details-content followed by a state pseudo-class");

  test(function() {
    assert_equals(CSS.supports("selector(::details-content:lang(en)"), true);
  }, "selector() function accepts ::details-content followed by :lang()");

  test(function() {
    assert_equals(CSS.supports("selector(::details-content:nth-of-type(slot)"), false);
  }, "selector() function doesn't accept ::details-content followed by a tree-structural pseudo-class");

  test(function() {
    assert_equals(CSS.supports("selector(::details-content:has(a)"), false);
  }, "selector() function doesn't accept ::details-content followed by :has()");

  test(function() {
    assert_equals(CSS.supports("selector(::details-content::part(a)"), false);
  }, "selector() function doesn't accept ::details-content followed by a part");
</script>
